Staging -> Main

+++ src/Core/Data/Dto/Transaction/CreateTransactionRequestDto.cs
@@ -4,7 +4,7 @@ namespace GamaEdtech.Data.Dto.Transaction
     {
         public int UserId { get; set; }
         public long? IdentifierId { get; set; }
-        public int Points { get; set; }
+        public long Points { get; set; }
         public string? Description { get; set; }
     }
 }
